Updated mutex sequences

+++ project5/db_project/db/include/trxes/LockContainer.h
@@ -9,13 +9,13 @@ private:
     std::map<std::size_t, LockEntry*>* keyMap{};
     pthread_mutex_t mutex{};
 
-    std::size_t getHashKey(int table_id, uint64_t pagenum);
+    std::size_t getHashKey(int64_t table_id, uint64_t pagenum);
 
 public:
     LockContainer();
     ~LockContainer();
 
-    LockEntry* getOrInsert(int table_id, uint64_t pagenum);
+    LockEntry* getOrInsert(int64_t table_id, uint64_t pagenum);
     int countEntries();
 };
